+++ plugin/demo/scenes/snapshots/Snapshots.gd
@@ -49,7 +49,7 @@ func _connect_save_game_data() -> void:
 	)
 	SnapshotsClient.game_loaded.connect(
 		func(snapshot: SnapshotsClient.Snapshot):
-			if snapshot == null:
+			if !snapshot:
 				print("Snapshot not found!")
 				return